include: remove note from `LIBOLED_SET_CONTRAST_CTRL`
The note was referencing that the command requires arguments. I need to
find a way to mark a command as requiring arguments instead of littering
notes everywhere.

Also, replace 'enum *_commands' with 'enum *_command'. Having the plural
of 'command' as the typename of the 'enum' felt a bit weird to see as
typenames don't usually end in the plural form of a word from what I can
remember.