docs&ssd1306.h: remove custom command `@cmdarg`
Documenting what the arguments for the commands are feels pointless when
that information is already in the datasheet. The functions I'm going to
end up writing will 'document' what the arguments for the commands are.

However, I'm going to keep '@cmdarg_or'. At first glance, it's not
apparent from the header file that the arguments must be or'd with the
command.