meson.build: set `default_library` to `static`
This library can be used on platforms that support dynamic libraries,
but, for the most part, it will be used on bare-metal. The default can
be changed when setting up the build directory for the first time.

    $ meson -Ddefault_library=shared "$BUILD_DIR"

or

    $ meson -Ddefault_library=both "$BUILD_DIR"

for both static and shared.

If the build directory was already setup then make sure to add the
--reconfigure option to meson.

    $ meson --reconfigure -Ddefault_library=shared "$BUILD_DIR"