README.md: build a TOC by explicitly defining anchors
Github has a way of automatically linking to parts of 'README's. This is
nice when building a table of contents. However, not all Markdown
engines support it the way Github does it. I got around this by
explicitly labelling the section headers as anchors and linking to them
using the usual idiom

    [blah](#defined_anchor)