platform.h: define a platform dependent "interface"
All functions provided by 'libSSD1306' will require a
'struct ssd1306_ctx *' as its first argument. The purpose of the struct
is to configure the method of communication between the mcu and the
'SSD1306'. Keeping track of all possible ways of configuring peripherals
doesn't sound maintainable. Instead, this 'interface' needs to be
implemented by the user and it should be enough. It's a pattern also
seen in the linux kernel.