.clang-format: break after return type on fn defs.
Breaking after the return type makes it easier to grep for function
defintions. For example,

    git grep ^ssd1306_

greps for all lines that begin with 'ssd1306_'. The only things that
begin with that pattern should be function names that start with
'ssd1306_'.

I also made it so the opening brace for function defintions are on their
own line. This makes the return type, function name, first line of the
body of the function less cluttered.

I also made it so wrapped function names, like breaking after the return
type, don't get indented.