README.md&platform.h: `const` the callbacks in `ctx`
A quick check on godbolt.com showed no difference in the generated
assembly between calling a function from a const/non-const function
pointer in multiple scenarios.

1. modifying a global int and returning the value
2. adding two to the parameter and returning it
3. simply returning a '2' (didn't expect much from this one)

Granted, these are simple examples and the compiler was probably able
to see through all of them, even without optimizations enabled.
I need to come up with a better test case to see if/when there's a
difference. It probably makes little to no difference in 'libSSD1306'.