platform.h: add a `width` and `height` field to `ctx`
I will need to know the bounds of the screen in order to properly
configure anything relating to the width/height of the OLED. The
'SSD1306' is 128x64, however, having a configurable width/height will
make it easier to extend the libraries reach to future versions of the
'SSDXXXX' line and clones.

Having the width and height be compile time constants would be nice but
that might be a pre-mature optimation. For now, I'll just keep it const.
It is something to keep in mind though!