ssd1306.h: remove `NUM_{COLS,PAGES}`
Instead of comparing against these values for boundary checking, use the
supplied 'width' and 'height' fields of 'struct ssd1306_ctx'. This will
allow for a more modular design. The values won't be compile time
constants but they are defined as 'const' in the 'struct' so it may not
be that bad.