ssd1306.h: properly define the `commands` module
note: The preferred way of calling `doxygen` commands is to prefix them
with `@` and not `\`. The reason why this commit message uses `\` is to
not `@` a user by accident.

Before, there was a parent `\defgroup` that grouped up multiple
`\defgroups` together, with `\{` and `\}`, in order to form the
`commands` module, i.e.

    \defgroup commands
    \{
        \defgroup fundamental_commands
        \{

        \}

        \defgroup scrolling_commands
        \{

        \}

        ...
    \}

This is incorrect. The correct way to form the `commands` module is to

1. define the parent group without grouping other things into it with `\{`
   and `\}`
2. explicitly `\ingroup` the parent group within each child `\defgroup`
   like so

    \defgroup commands

    \defgroup fundamental_commands
    \ingroup commands
    \{

    \}

    \defgroup scrolling_commands
    \ingroup commands
    \{

    \}

    ...