ssd1306.h: only use one `\note` for every comment
There were some comments that had multiple `\note`s in the hopes that it
would produce a better layout in the generated documentation. This
worked, but, this would produce inconsistent output in other
documentation generators. `doxygen` merges multiple `\note`s within the
same comment but `Sphinx` and `breathe` don't. It might be hidden behind
some option, but, it wasn't a clean solution in the first place.