README.md: use `struct::field` syntax for linking
Since the 'README' is being used as the main page for the 'doxygen'
documentation, the 'struct::field' notation produces the correct links.

It also doesn't hurt getting into the habit of doing it like that in all
places. I think most, if not all, either know or could easily figure the
notation out.