{include,src}/: define `send_{cmd,data}` wrappers
It can be tedious to explicitly call the field of a struct and pass the
struct to it along with the other arguments to the function. This commit
also updates usages of calling 'ssd1306_ctx::send_cmd' directly with the
wrapper function.

I also added a note about the wrapper functions on the callback fields
of 'ssd1306_ctx'.